requirements.txt
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.6
redis==5.1.1
six==1.16.0
requests==2.32.3
sniffio==1.3.1
PyJWT==2.9.0
sqlparse==0.5.1
stripe==11.1.0
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
urllib3==2.2.3
flake8==7.1.1
drf-spectacular==0.27.2
